begin adding in Numba tutorial and methods
I don't think there's much utility in teaching weave at this point.
It's 10 years old and Numba is getting faster every day (more or less).
If we decide to keep the SRJ stuff, I think it makes sense to split it
off into a separate notebook, as the Numba intro is going to make this
one quite long.

Downside:  need to reprogram the entire multigrid lesson to stop using
Weave.  so be it