add working directory with gitignore
As per conversation with @labarba, this adds a "working" directory for
students to save their own notebooks in and avoids merge conflicts when
they have to update their forks.  As an added bonus, there's a gitignore
file in the working directory that will stop git from tracking the ipynb
checkpoint files.