fix header formatting throughout modules 1-4
Recent Jupyter update seems to have changed Markdown behavior regarding
headline formatting.

Previously, a header was defined in markdown by using a variable number
of # signs followed by the text of the header.

In the new version, there must be a space separating the header text
from the # signs or it doesn't render