deep editing pass on conjugate gradient lesson
* add relative links to all mentions of previous lessons
* format references to use markdown links
* add html anchor link to reference section for repeated Shewchuk
references in text
* few variable name changes to increase consistency with written
explanations

For both steepest descent and conjugate gradients, I've expanded the
explanation of the methods.  I provided a little more background on what
the residual is (moved from the multigrid notebooks, since students will
see this one first).  Also added more details on the individual steps of
both algorithms, things that were implicitly available via the code at
hand, but not explicitly written out.

I'm trying to avoid just writing Shewchuk's entire paper in here in
notebook form, but I might add in a few of the derivations as an
appendix.  It would be great if students go and read Shewchuk and work
through it, but it's 50 pages and I am not an optimist.  Even without a
full explanation of every part of CG, I might throw in a few simpler
parts of it -- e.g. the formula for $r^{n+1}$.