Changed 'centrifugal acceleration' to 'centripetal acceleration' in 'Phugoid Theory'
In any circular motion of a body, the body continuously falls accelerating 'towards center' which we call 'centripetal acceleration'. This 'centrepetal acceleration' makes the body feel a fictitious force acting 'away from the center' which we call 'centrifugal force'.

Thus I think what we encounter in a typical circular motion is 'centripetal aceleration' as opposed to 'centrifugal acceleration'