add FBD and change force balance explanation
As per discussion with @bknaepen and @ncclementi about the inherent
_UGH_ feelings around "centifugal acceleration" and the general
wonkiness of the whole thing, this is instead a reworked explanation
that derives the same equation from a simple free body diagram.

References to a moving frame have been removed and the FBD and the
previosu diagram are now SVGs so that they can be edited in the future
without loss of fidelity.

Please take a look and advise if anyone would like any further changes /
tweaks made.  I think this is the right direction to go with this, but I
want to also make sure that this is FINISHED.

Awesome!  Thanks to all!